\name{DPEEM-package}
\alias{DPEEM-package}
\alias{DPEEM}
\docType{package}
\title{
\packageTitle{DPEEM}
}
\description{
\packageDescription{DPEEM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DPEEM}
\packageIndices{DPEEM}
just library it and a GUI can be generate, or you can use functions as below:

funIntegrate(workdir = getwd(),
                        outputdir = workdir,
                        FDMax, FDMin,
                        ROIfilename1,
                        DA1_to_csv = T,
                        preprocess = T,
                        TTFFD = T,
                        FTnum = T,
                        GazeDuration = T,
                        Regression = T,
                        SaccadeLength = T,
                        SecondPassTime = T,
                        FixationProportion = T,
                        DataIntegrate = T)
}
\author{
\packageAuthor{DPEEM}

Maintainer: \packageMaintainer{DPEEM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

